+++ tests/testthat/test_nest.R
@@ -9,6 +9,7 @@ test_all(strict_args = list(exclude_from_man = c("decorate_grob.Rd", #partial ar
                                                  "t_coxph_pairwise.Rd",
                                                  "t_el_disposition.Rd",
                                                  "t_el_forest_tte.Rd",
+                                                 "t_el_odds_ratio.Rd",
                                                  "t_forest_rsp.Rd",
                                                  "t_forest_tte.Rd",
                                                  "t_km.Rd",